requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
django-crispy-forms==1.7.2
idna==2.7
nltk==3.3
numpy==1.14.5
protobuf==3.6.0
pycparser==2.18
PyMySQL==0.9.2
PyYAML==3.13
six==1.11.0
textblob==0.15.1
vaderSentiment==3.2.1
virtualenv==16.0.0
virtualenvwrapper-win==1.2.5
